Testing a refactor of Activity dependent Bluetooth ++
